+++ indico/modules/categories/client/js/calendar.js
@@ -105,7 +105,10 @@ import CalendarLegend from './components/CalendarLegend';
       },
       events({start, end}, successCallback, failureCallback) {
         function updateCalendar(data) {
-          const attr = groupBy === 'category' ? 'categoryId' : 'venueId';
+          const attr = {category: 'categoryId', location: 'venueId', room: 'roomId'}[groupBy];
+          if (!attr) {
+            throw new Error(`Invalid "groupBy": ${groupBy}`);
+          }
           const filteredEvents = data.events.filter(e => !filteredLegendElements.has(e[attr] ?? 0));
           successCallback(filteredEvents);
           const toolbarGroup = $(containerCalendarSelector).find(
@@ -145,27 +148,67 @@ import CalendarLegend from './components/CalendarLegend';
             filteredLegendElements.clear();
             calendar.refetchEvents();
           };
-          const onElementSelected = (id, checked) => {
-            if (checked) {
-              filteredLegendElements.delete(id);
-            } else {
-              filteredLegendElements.add(id);
+          const onElementSelected = (elementId, checked, elementSubitems, parent) => {
+            const handleSingle = (singleId, singleChecked) => {
+              if (singleChecked) {
+                filteredLegendElements.delete(singleId);
+              } else {
+                filteredLegendElements.add(singleId);
+              }
+            };
+            const handleRecursive = (recursiveId, recursiveSubitems) => {
+              handleSingle(recursiveId, checked);
+              recursiveSubitems.forEach(({id, subitems}) => handleRecursive(id, subitems));
+            };
+
+            // handle descendents
+            handleRecursive(elementId, elementSubitems);
+
+            // handle parent
+            if (parent) {
+              const parentChecked = !parent.subitems.every(({id}) =>
+                filteredLegendElements.has(id)
+              );
+              handleSingle(parent.id, parentChecked);
             }
+
             calendar.refetchEvents();
           };
-          items = items.map(item => ({...item, checked: !filteredLegendElements.has(item.id)}));
+          const selectAll = () => {
+            items.forEach(({id, subitems}) => onElementSelected(id, true, subitems));
+            calendar.refetchEvents();
+          };
+          const deselectAll = () => {
+            items.forEach(({id, subitems}) => onElementSelected(id, false, subitems));
+            calendar.refetchEvents();
+          };
+          const filterLegendElement = item => {
+            item.checked = !filteredLegendElements.has(item.id);
+            item.subitems.forEach(filterLegendElement);
+          };
+          items.forEach(filterLegendElement);
           ReactDOM.render(
             <CalendarLegend
               items={items}
               groupBy={data.group_by}
               onFilterChanged={onFilterChanged}
               onElementSelected={onElementSelected}
+              selectAll={selectAll}
+              deselectAll={deselectAll}
             />,
             legendContainer
           );
         }
 
-        function setupLegendByAttribute(events, items, attr, defaultTitle, rootId, rootTitle) {
+        function setupLegendByAttribute({
+          events,
+          items,
+          attr,
+          defaultTitle,
+          rootId,
+          rootTitle = undefined,
+          sortMethod = undefined,
+        }) {
           const itemMap = items.reduce(
             (acc, {id, title, url}) => ({
               ...acc,
@@ -191,6 +234,8 @@ import CalendarLegend from './components/CalendarLegend';
                   url: item.url,
                   isSpecial,
                   id,
+                  subitems: [],
+                  parent: null,
                 },
               ];
             }, [])
@@ -200,31 +245,91 @@ import CalendarLegend from './components/CalendarLegend';
               } else if (b.isSpecial) {
                 return 1;
               }
-              return a.title.localeCompare(b.title);
+              return sortMethod ? sortMethod(a.title, b.title) : a.title.localeCompare(b.title);
             });
+        }
+
+        function setupLegendByRoom(events, locations, rooms) {
+          const collator = new Intl.Collator(undefined, {numeric: true, sensitivity: 'base'});
+          const result = setupLegendByAttribute({
+            events,
+            items: rooms,
+            attr: 'roomId',
+            defaultTitle: Translate.string('No room'),
+            rootId: 0,
+            rootTitle: Translate.string('No room'),
+            sortMethod: collator.compare,
+          });
+          // if there are no locations no need to indent
+          if (locations.length <= 1) {
+            return result;
+          }
+          const roomMap = rooms.reduce(
+            (acc, value) => ({
+              ...acc,
+              [value.id]: value,
+            }),
+            {}
+          );
+          const locationMap = locations.reduce(
+            (acc, value) => ({
+              ...acc,
+              [value.id]: value,
+            }),
+            {}
+          );
+          const groupedLocations = {};
+          const noRoom = [];
+          result.forEach(item => {
+            const room = roomMap[item.id];
+            if (!room) {
+              // this is the "No room" case
+              noRoom.push(item);
+            } else {
+              const location = locationMap[room.venueId];
+              if (!groupedLocations[location.id]) {
+                groupedLocations[location.id] = {
+                  // we make the ID negative in order not to collide with room IDs
+                  id: -location.id,
+                  title: location.title,
+                  isSpecial: false,
+                  subitems: [],
+                  parent: null,
+                };
+              }
+              groupedLocations[location.id].subitems.push({
+                ...item,
+                parent: groupedLocations[location.id],
+              });
+            }
+          });
+          return [...noRoom, ...Object.values(groupedLocations)];
         }
 
         function updateLegend(data) {
           let items;
           switch (data.group_by) {
             case 'category':
-              items = setupLegendByAttribute(
-                data.events,
-                data.categories,
-                'categoryId',
-                Translate.string('No category'),
-                categoryId,
-                Translate.string('This category')
-              );
+              items = setupLegendByAttribute({
+                events: data.events,
+                items: data.categories,
+                attr: 'categoryId',
+                defaultTitle: Translate.string('No category'),
+                rootId: categoryId,
+                rootTitle: Translate.string('This category'),
+              });
               break;
             case 'location':
-              items = setupLegendByAttribute(
-                data.events,
-                data.locations,
-                'venueId',
-                Translate.string('No venue'),
-                0
-              );
+              items = setupLegendByAttribute({
+                events: data.events,
+                items: data.locations,
+                attr: 'venueId',
+                defaultTitle: Translate.string('No venue'),
+                rootId: 0,
+              });
+              break;
+            case 'room':
+              items = setupLegendByRoom(data.events, data.locations, data.rooms);
               break;
             default:
               items = [];